BasicLinearAlgebra

Description

Documentation

newtype Matrix

Type of matrix contains a list of lists of Int

Constructors

Matrix [[Rational]]

extractM :: Matrix -> [[Rational]]

Used to extract the list of lists of Int from Matrix type

extractMaybe :: Maybe a -> a -> a

Extract the data from Maybe Monad | if value empty it returns the fallback value | the second argument is a fallback value

mSize :: Matrix -> Maybe (Line, Column)

Get the size of the matrix in a tuple (Int, Int)

mSizeMult :: Matrix -> Matrix -> Maybe (Line, Column)

Returns the resulting size of the matrix after multiplication of m and m'.

mIsSquare :: Matrix -> Bool

Verify if matrix is square matrix

mSizeEq :: Matrix -> Matrix -> Bool

Verify if both matrix are of equal size

mMultipliable :: Matrix -> Matrix -> Bool

Verify if both matrices can be multiplied

mMultiply :: Matrix -> Matrix -> Matrix

Multiply both matrices (must first be verified with mMultipliable)

mAdd :: Matrix -> Matrix -> Matrix

Add matrices together (must first be verified with mSizeEq)

mScale :: Matrix -> Scale -> Matrix

Apply a scale to a matrix